feat: user delete (#28)

+++ src/main/java/ftn/userservice/services/UserService.java
@@ -1,11 +1,14 @@
 package ftn.userservice.services;
 
 import ftn.userservice.domain.dtos.ChangePasswordRequest;
+import ftn.userservice.domain.dtos.ReservationDto;
 import ftn.userservice.domain.dtos.UserDto;
 import ftn.userservice.domain.dtos.UserUpdateRequest;
+import ftn.userservice.domain.entities.Role;
 import ftn.userservice.domain.entities.User;
 import ftn.userservice.domain.mappers.UserMapper;
 import ftn.userservice.exception.exceptions.BadRequestException;
+import ftn.userservice.exception.exceptions.ForbiddenException;
 import ftn.userservice.exception.exceptions.NotFoundException;
 import ftn.userservice.repositories.UserRepository;
 import ftn.userservice.utils.AuthUtils;
@@ -13,6 +16,10 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -22,6 +29,7 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final RestService restService;
 
     public UserDto getById(UUID id) {
         User user = userRepository.findById(id).orElseThrow(() -> new NotFoundException("User doesn't exist"));
@@ -68,7 +76,32 @@ public class UserService {
 
     public void delete() {
         UUID id = AuthUtils.getLoggedUserId();
-        //TODO check reservations when other services are done
+        User user = userRepository.getReferenceById(id);
+        List<ReservationDto> reservations = new ArrayList<>();
+        boolean found = false;
+
+        if (user.getRole() == Role.GUEST) {
+            reservations = restService.getGuestReservations();
+        } else if (user.getRole() == Role.HOST) {
+            reservations = restService.getHostReservations();
+        }
+
+        for (ReservationDto reservation: reservations) {
+            if (reservation.getStatus() == ReservationDto.ReservationStatus.ACTIVE) {
+                if (reservation.getDateFrom().isAfter(LocalDateTime.now().minusDays(1))) {
+                    found = true;
+                    break;
+                }
+            }
+        }
+
+        if (found) {
+            throw new ForbiddenException("Can't delete account with reservations");
+        }
+
+        if (user.getRole() == Role.HOST) {
+            restService.deleteHostLodges(id);
+        }
         userRepository.deleteById(id);
     }
